Move common fixtures to first class module
This introduces ansible_base.lib.testing and moves some of the
more useful common fixtures out to that so other apps can use them,
too.
